% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{armToContrast}
\alias{armToContrast}
\title{transform arm data to contrast data if needed}
\usage{
armToContrast(sheet, options)
}
\arguments{
\item{sheet}{df, a dataframe derived from a sheet of the uploaded excel book}

\item{option}{list, a list of arguments selected in the ui}
}
\description{
transform arm data to contrast data if needed
}
