% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{disableOption}
\alias{disableOption}
\title{disable an option within a radio group}
\usage{
disableOption(value, id)
}
\arguments{
\item{value}{string, the value of the option}

\item{id}{string, the id (not selector) of the option group}
}
\value{
void
}
\description{
disable an option within a radio group
}
