% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{getFormattedData}
\alias{getFormattedData}
\title{format data to be passed to plotting function}
\usage{
getFormattedData(options, sheets, input)
}
\arguments{
\item{options}{list, list keyed by sheet of arguments to pass to functions}

\item{sheets}{list, list keyed by sheet name of user-supplied dataframes}

\item{input}{reactiveValues, the list of inputs, which will contain formatting values}
}
\value{
dataframe, dataframe that can be passed to rank heat plot function if successful,
else a list of fields that failed the connectivity test
}
\description{
format data to be passed to plotting function
}
