% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{getMissingOptions}
\alias{getMissingOptions}
\title{Determine whether the given set of options is complete}
\usage{
getMissingOptions(optionSet)
}
\arguments{
\item{optionSet}{list, the list of set options, keyed by sheet name}
}
\value{
list, a list of missing fields, keyed by sheet name
}
\description{
Determine whether the given set of options is complete
}
