% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{getNetmeta}
\alias{getNetmeta}
\title{Run netmeta function with user values converting arm to contrast}
\usage{
getNetmeta(options, sheet)
}
\arguments{
\item{options}{list, user-provided arguments to pass to netmeta}

\item{sheet}{dataframe, data from a user sheet}
}
\value{
dataframe, dataframe that can be passed to rank heat plot function
}
\description{
Run netmeta function with user values converting arm to contrast
}
