% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{getRanks}
\alias{getRanks}
\title{get rankings}
\usage{
getRanks(options, netmetaRes, sheetName)
}
\arguments{
\item{options}{list, user-provided arguments to pass to netmeta}

\item{netmetaRes}{netmeta object, results of netmeta function}

\item{sheetName}{char, the name of the sheet}
}
\value{
dataframe, dataframe that can be passed to rank heat plot function
}
\description{
get rankings
}
