% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankHeatCircos.R
\name{rhp.rankheatplotCircos}
\alias{rhp.rankheatplotCircos}
\title{Plot the rank-heat graphic}
\usage{
rhp.rankheatplotCircos(
  chartData,
  format = "percentage",
  cexLabel = 0.65,
  cexValue = 0.75,
  cexSector = 1
)
}
\arguments{
\item{chartData, }{datagram}

\item{format, }{chr}

\item{cexLabel, }{float, scale of label text}

\item{cexValue, }{float, scale of value text}

\item{cexSector, }{float, scale of sector text}
}
\value{
void
}
\description{
Plot the rank-heat graphic
}
