% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{validateSheet}
\alias{validateSheet}
\title{Confirm that the sheet has the appropriate columns, types, and contents
for the options selected}
\usage{
validateSheet(options, sheet)
}
\arguments{
\item{options}{list, list of selected options}

\item{sheet}{dataframe, a user-submitted sheet}
}
\value{
list, a list of errors keyed by field name
}
\description{
Confirm that the sheet has the appropriate columns, types, and contents
for the options selected
}
